# src/main/resources/openapi.yaml
openapi: 3.0.4
info:
  title: School API
  version: 1.0.0
#  description: API para gestionar colegio
  description: |-
    API gestion school, this API is a simple example of how to use OpenAPI 3.0 with Spring Boot.

    Some useful links:
    - [Developer LinkedIn](https://www.linkedin.com)
    - [Repository GitHub](https://github.com/josephmn/msv-maven-student/blob/feature/issue0001/src/main/resources/openapi.yaml)
  termsOfService: https://swagger.io/terms/
  contact:
    email: developer.senior@example.pe
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:8082
tags:
  - name: student
    description: Operations about student
  - name: teacher
    description: Operations about teacher
paths:
  /students:
    get:
      tags:
        - student
      summary: Get all student.
      description: Get all student status.
      operationId: getAllStudents
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ResponseStudentDto'
        '400':
          description: Invalid username supplied
        '404':
          description: User not found
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    post:
      tags:
        - student
      summary: Create student.
      description: Create student in database.
      operationId: createStudent
      requestBody:
        description: Created user object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestStudentDto'
        required: true
      responses:
        '201':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseStudentDto'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /students/actives:
    get:
      tags:
        - student
      summary: Get all student actives.
      description: Get all student status actives.
      operationId: getAllStudentsActives
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ResponseStudentDto'
        '400':
          description: Invalid username supplied
        '404':
          description: User not found
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /students/{id}:
    get:
      tags:
        - student
      summary: Get student by ID.
      description: Get student by ID.
      operationId: getStudentById
      parameters:
        - name: id
          in: path
          required: true
          description: ID of student
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseStudentDto'
        '400':
          description: Invalid ID supplied
        '404':
          description: User not found
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      tags:
        - student
      summary: Put student by ID.
      description: Put student by ID.
      operationId: updateStudentById
      parameters:
        - name: id
          in: path
          required: true
          description: ID of student
          schema:
            type: integer
            format: int64
      requestBody:
        description: Created user object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestStudentDto'
        required: true
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseStudentDto'
        '400':
          description: Invalid ID supplied
        '404':
          description: User not found
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      tags:
        - student
      summary: Delete student by ID.
      description: Delete student by ID.
      operationId: deleteStudentById
      parameters:
        - name: id
          in: path
          required: true
          description: ID of student
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
        '400':
          description: Invalid ID supplied
        '404':
          description: User not found
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /teachers:
    get:
      tags:
        - teacher
      summary: Get all teacher.
      description: Get all teacher.
      operationId: getAllTeachers
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ResponseTeacherDto'
        '400':
          description: Invalid username supplied
        '404':
          description: User not found
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    post:
      tags:
        - teacher
      summary: Create teacher.
      description: Create teacher in database.
      operationId: createTeacher
      requestBody:
        description: Created teacher object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestTeacherDto'
        required: true
      responses:
        '201':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseTeacherDto'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

components:
  schemas:
    RequestStudentDto:
      type: object
      properties:
        document:
          type: string
          example: 78653423
        name:
          type: string
          example: Jhon
        lastName:
          type: string
          example: James
        status:
          type: boolean
          example: true
        age:
          type: integer
          format: int32
          example: 30

    ResponseStudentDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        document:
          type: string
          example: 78653423
        name:
          type: string
          example: Jhon
        lastName:
          type: string
          example: James
        status:
          type: boolean
          example: true
        age:
          type: integer
          format: int32
          example: 30

    RequestTeacherDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        document:
          type: string
          example: 45453245
        name:
          type: string
          example: John
        lastName:
          type: string
          example: James
        status:
          type: boolean
          example: true
        age:
          type: integer
          format: int32
          example: 30

    ResponseTeacherDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        document:
          type: string
          example: 45453245
        name:
          type: string
          example: John
        lastName:
          type: string
          example: James
        status:
          type: boolean
          example: true
        age:
          type: integer
          format: int32
          example: 30

    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required:
        - code
        - message